Feat/authentication (#11)
* feat: login page and basic request for JWT token

* chore: gitignore .idea directory generated by IntelliJ IDEA

* chore: install jose for JWT decoding

* feat: /logout for deleting JWT cookie and redirecting back to main page

* feat: decode jwt token into user data

* style: format code

* feat: basic login and registeration functionality

User logs in with username and password and gets a JWT token in return 
which is used automatically in subsequent requests to authenticate the 
user.

* chore(docker): simplified dockerfile and docker-compose; changed port to 3000

* chore(docker): configure container network

Frontend container now runs on the same network with backend. This
allows, coupled with API_ENDPOINT environment variable and `handleFetch`
method in `./src/hooks.server.js`, both frontend and backend containers
to run at the same time. Frontend development locally remains
unaffected.